# contracts/openapi/v1/openapi.json
{
  "components": {
...
    "parameters": {
      "CorrelationId": {
        "description": "UUID for a multi-step operation. A valid value is preserved; when absent the runtime generates one.",
        "in": "header",
        "name": "X-Correlation-ID",
        "required": false,
        "schema": {
          "format": "uuid",
          "type": "string"
        }
      }
    },
...
    "securitySchemes": {
      "environmentAccess": {
        "bearerFormat": "JWT",
        "description": "Environment-level operator or service identity. It is not a product user session or product authorization contract.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Foundation operational contract. Product resources will be added below /api/v1 without changing these side-effect-free probes.",
    "title": "Umbral Runtime API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/agent/ops/overview": {
      "get": {
        "operationId": "agentOpsOverview",
        "parameters": [
          {
            "$ref": "#/components/parameters/CorrelationId"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "summary": "Agent Ops Overview",
        "tags": [
          "agent-ops"
        ]
      }
    },
...
  },
  "security": [
    {
      "environmentAccess": []
    }
  ],
  "servers": [
    {
      "description": "Local API",
      "url": "http://127.0.0.1:8000"
    }
  ],
...
}